lesson learnt
don't use -n and -i at the same time with sed -- one of them silences
the output and one of them writes the output to the input file (making
backups with the provided suffix, if necessary). in my case, there was
no output to speak of so the whole operation resulted in a completely
empty file.