add functions for setting up emacs
- setup_emacs() checks if either of the emacs directories exist. if they
do, it is assumed that an installation already exists and therefore the
script exits with a non-zero status.

- if the directories don't exist, they are created, doom emacs repo is
cloned and then installed. thereafter, for the doom binary to be
available to be called directly, it is linked in a directory already
present in $PATH which is ${HOME}/.local/bin.

- finally, the configuration is linked and then doom is synced to the
new state.